requirements.txt
kafka-python==2.0.2
pyspark==3.3.2
psycopg2-binary==2.9.5
pandas==2.0.0
numpy==1.24.2
matplotlib==3.7.1